rename file, increaes key collection size
